+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.7.0] - 2017-11-03
+### Added
+- DF-1191 Added attachment support for email services and email templates
+- Upgrade Swagger to OpenAPI 3.0 specification
+- Fix SMTP port casting as integer
+
 ## [0.6.0] - 2017-08-17
 ### Changed
 - Reworking API doc usage and generation
@@ -33,7 +39,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-11-30
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-email/compare/0.6.0...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-email/compare/0.7.0...HEAD
+[0.7.0]: https://github.com/dreamfactorysoftware/df-email/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/dreamfactorysoftware/df-email/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/dreamfactorysoftware/df-email/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/dreamfactorysoftware/df-email/compare/0.3.0...0.4.0